requirements.txt
blinker==1.8.2
blockfrost-python==0.5.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.2
